% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFrac.R
\name{calculateVarFrac}
\alias{calculateVarFrac}
\title{Calculate the fraction of phenotypic variance explained by as set of 
significant genetic effect and/or interaction effects.}
\usage{
calculateVarFrac(std_bG, std_bI, matcor, varY, type)
}
\arguments{
\item{std_bG}{is the vector of standardized genetic effect sizes}

\item{std_bI}{is the vector of standardized interaction effect sizes}

\item{matcor}{is the genotype correlation matrix}

\item{varY}{is the phenotypic variance in the pooled sample}

\item{type}{indicates wether only genetic or interactions effects should be 
considered or if both should be considered jointly.}
}
\value{
The fraction of phenotypic variance explained by user-specified effects.
}
\description{
This version does not take into account potential biases in neither the 
estimation of effect sizes nor the estimation of the correlation matrix
}
\examples{
std_bG <- rnorm(5, 0, 0.01)
std_bI <- rnorm(5, 0, 0.001)
matcor <- cor(matrix(runif(5*5, 1, 5), nrow = 5))
varY <- 2.25
calculateVarFrac(std_bG, std_bI, matcor, varY, "G")
calculateVarFrac(std_bG, std_bI, matcor, varY, "I")
calculateVarFrac(std_bG, std_bI, matcor, varY, "J")

}
