% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.R
\name{getAdditivelyCodedMatrix}
\alias{getAdditivelyCodedMatrix}
\title{Code additively the genotypes}
\usage{
getAdditivelyCodedMatrix(genoMat, lind)
}
\arguments{
\item{genoMat}{is the matrix of genotype with two colums per individual
Rows are variants and columns are individuals}

\item{lind}{is the list of individual in the genotype matrix}
}
\value{
The additively coded genotype matrix
}
\description{
Sum the two columns per individual of the input matrix (each corresponding 
to one allele) to get the additive coding of the individual at the locus.
}
