% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWAS}
\alias{GWAS}
\title{Example of GWAS result input file}
\format{An object of class \code{data.frame} with 15 rows and 13 columns.}
\usage{
GWAS
}
\description{
Example of GWAS result input file
}
\author{
Vincent Laville \email{vincent.laville@pasteur.fr}
}
\keyword{data}
